Intake was rewritten in GamePieceManipulator to use a state machine
RollerIntakeMechanism is a new class that handles the
intake state machine. It was made into a seperate class
that GamePiecemanipulator has an object if to improve
code modularity in case we
switch from a roller design to a vacuum design.

GamePieceManipulator is also in the middle of being
rewritten to use a state machine for deploy. It is
not finished.

Added many joysticks buttons to Hardware for use
in Teleop when calling GamePieceManipulator in teleop

Minor changes to Teleop for testing code